move spawn and contact events to systems::events submodule
